ai_financial_advisor/requirements.txt
crewai
crewai[tools]
pandas
numpy
matplotlib
python-dotenv
requests